+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changelog for NeoFS Node
 ### Added
 - Embedded Neo contracts in `contracts` dir (#2391)
 - `dump-names` command for adm
+- `renew-domain` command for adm
 
 ### Fixed